pyproject.toml
[tool.poetry]
name = "piplayer"
version = "0.1.0"
description = "A command line interface to help set up a raspberry pi as a video player."
authors = ["Sam Lavigne <splavigne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
piplayer = "piplayer:cli"
